Refactored `pages` to `containers`
