+++ client/src/app/containers/PointOfSale.js
